+++ aws/container-linux/kubernetes/outputs.tf
@@ -37,6 +37,11 @@ output "kubeconfig" {
 
 # Outputs for custom load balancing
 
+output "nlb_id" {
+  description = "ARN of the Network Load Balancer"
+  value       = "${aws_lb.nlb.id}"
+}
+
 output "worker_target_group_http" {
   description = "ARN of a target group of workers for HTTP traffic"
   value       = "${module.workers.target_group_http}"