+++ aws/container-linux/kubernetes/variables.tf
@@ -71,6 +71,12 @@ variable "worker_price" {
   description = "Spot price in USD for autoscaling group spot instances. Leave as default empty string for autoscaling group to use on-demand instances. Note, switching in-place from spot to on-demand is not possible: https://github.com/terraform-providers/terraform-provider-aws/issues/4320"
 }
 
+variable "worker_target_groups" {
+  type        = "list"
+  description = "Additional target group ARNs to which worker instances should be added"
+  default     = []
+}
+
 variable "controller_clc_snippets" {
   type        = "list"
   description = "Controller Container Linux Config snippets"