+++ aws/container-linux/kubernetes/workers/workers.tf
@@ -19,6 +19,7 @@ resource "aws_autoscaling_group" "workers" {
   target_group_arns = [
     "${aws_lb_target_group.workers-http.id}",
     "${aws_lb_target_group.workers-https.id}",
+    "${var.target_groups}",
   ]
 
   lifecycle {